+++ src/kingdom/storage.ts
@@ -1,36 +1,8 @@
-import {FameType, getCultEventMilestones, Kingdom} from './data/kingdom';
+import {Kingdom} from './data/kingdom';
 import {groupBy} from '../utils';
 
 export function getKingdom(sheetActor: Actor): Kingdom {
     const kingdom = sheetActor.getFlag('pf2e-kingmaker-tools', 'kingdom-sheet') as Kingdom;
-    // migrations
-    if (typeof kingdom.fame === 'number' && 'fameNext' in kingdom && 'fameType' in kingdom) {
-        const current = kingdom as Kingdom & {fame: number, fameNext: number, fameType: FameType};
-        kingdom.fame = {
-            now: current.fame,
-            next: current?.fameNext ?? 0,
-            type: current?.fameType ?? 'famous',
-        };
-    }
-    if (!kingdom.milestones.some(milestone => milestone.name.startsWith('Cult Event'))) {
-        getCultEventMilestones().forEach(milestone => kingdom.milestones.push(milestone));
-    }
-    if (kingdom.supernaturalSolutions === undefined) {
-        kingdom.supernaturalSolutions = 0;
-    }
-    if (kingdom.turnsWithoutCultEvent === undefined) {
-        kingdom.turnsWithoutCultEvent = 0;
-    }
-    if (kingdom.creativeSolutions === undefined) {
-        kingdom.creativeSolutions = 0;
-    }
-    if (kingdom.modifiers === undefined) {
-        kingdom.modifiers = [];
-    }
-    if (kingdom.settlements === undefined) {
-        kingdom.settlements = [];
-        kingdom.activeSettlement = '';
-    }
     console.log(kingdom);
     return kingdom;
 }