+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.23] - 2023-07-19
 
+### Added
+
+* Added migrations
+
 ### Fixed
 
-* Fix settlement levels randomly turning into strings; to fix this permanently, you need to edit a settlement and click
-  on save
+* Fix settlement levels randomly turning into strings; to fix this permanently
 
 ## [0.2.22] - 2023-07-14